lib: Rename geda_fill_type.[ch] functions.
